pyproject.toml
[tool.poetry]
name = "spotify-playlist-maker"
version = "0.1.0"
description = ""
authors = ["Matt Roberts <matt@antigeni.us>"]

[tool.poetry.dependencies]
python = "^3.10"
spotipy = "^2.23.0"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.22"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
